feat(it02|r4-hu02): Implementación de proxy precios
Se ha implementado el patrón proxy para evitar realizar peticiones costosas e innecesarias para obtener los precios de los combustibles (se actualiza cada 30 min) y de la luz (se actualiza diariamente).

Además el cálculo del coste se beneficia del patrón Strategy para evitar código repetido.

Para obtener un precio es recomendable acceder desde ServicioAPI, que hace de fachada para hacer más cómodo el acceso a los precios de forma eficiente

+++ app/src/main/java/es/uji/smallaris/model/IBuilderRutas.kt
@@ -10,7 +10,7 @@ interface IBuilderRutas {
     fun setTrayecto(trayecto: LineString): IBuilderRutas
     fun setDistancia(distancia: Float): IBuilderRutas
     fun setDuracion(duracion: Float): IBuilderRutas
-    fun setCoste(coste: Float): IBuilderRutas
+    fun setCoste(coste: Double): IBuilderRutas
     fun reset()
     fun getRuta(): Ruta
 }
